requirements.txt
streamlit
pandas
matplotlib
scikit-learn
joblib
numpy
langchain
langchain-google-genai
langchain-community
langgraph
chromadb
duckduckgo-search
python-dotenv
pydantic